:recycle: refactor: CategoriaController

+++ app/Http/Controllers/CategoriaController.php
@@ -37,9 +37,13 @@ class CategoriaController extends Controller
     public function store(CategoriaRequest $request)
     {
         $data = $request->all();
-        $this->categorias->create($data);
+        if (!empty($data['categoria'])) {
+            $this->categorias->create($data);
 
-        return redirect()->route('categoria.index')->with('success', 'Categoria criada com sucesso!');
+            return redirect()->route('categoria.index')->with('success', 'Categoria criada com sucesso!');
+        } else {
+            return redirect()->route('categoria.create')->with('danger', 'Categoria inválida!');
+        }
     }
 
     public function show($id)
@@ -51,27 +55,39 @@ class CategoriaController extends Controller
     public function edit($id)
     {
         $categoria = $this->categorias->find($id);
-        return view('categoria.crud', compact('categoria'));
+        if (!empty($categoria)) {
+            return view('categoria.crud', compact('categoria'));
+        } else {
+            return redirect()->route('categoria.index')->with('danger', 'Categoria não encontrada!');
+        }
     }
 
     public function update(CategoriaRequest $request, $id)
     {
         $data = $request->all();
         $categoria = $this->categorias->find($id);
-        $categoria->update($data);
+        if (!empty($categoria)) {
+            $categoria->update($data);
 
-        return redirect()->route('categoria.index')->with('success', 'Categoria alterada com sucesso!');
+            return redirect()->route('categoria.index')->with('success', 'Categoria alterada com sucesso!');
+        } else {
+            return redirect()->route('categoria.index')->with('danger', 'Categoria não encontrada!');
+        }
     }
 
     public function destroy($id)
     {
         $categoria = $this->categorias->find($id);
-        $produtos = $this->produtos->where('categoria_id', $id)->whereNotNull('imagem')->get();
-        foreach($produtos as $produto){
+        if (!empty($categoria)) {
+            $produtos = $this->produtos->where('categoria_id', $id)->whereNotNull('imagem')->get();
+            foreach ($produtos as $produto) {
                 Storage::disk('public')->delete(str_replace('/storage/', '', $produto->imagem));
-        }
-        $categoria->delete();
+            }
+            $categoria->delete();
 
-        return redirect()->route('categoria.index')->with('success', 'Categoria deletada com sucesso!');
+            return redirect()->route('categoria.index')->with('success', 'Categoria deletada com sucesso!');
+        } else {
+            return redirect()->route('categoria.index')->with('danger', 'Categoria não encontrada!');
+        }
     }
 }
